Add caching to CSV generation code.
A Catalyst client had trouble with browser timeouts while attempting to
get a questionnaire report. Execution time for the report was 132s on my
development machine. This patch adds MUC caching of the
questionnaire_quest_choice query results in response_select (most of the
savings) and folds the retrieval of user info in generate_csv into the
preceding query.

In the client's case, it reduced execution time to 23s (an 82% saving).

Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>