Properly clean up deleted questions.
The questionnaire module does not completely clean up a deleted
question at the moment. It deletes some data (responses and
attempts, for example) but leaves behind the question record
(marked as deleted) and choices in the case of multichoice qns.

One Catalyst client had a course containing a questionnaire
that took 35 minutes to restore. It was discovered that this
questionnaire had multiple deleted copies of a multichoice
question listing every suburb in Australia. Adding this code
caused the deleted questions to be properly cleaned up and
was part of a series of patches that reduced the restore time
to approximately 2 minutes.

Signed-off-by: Nigel Cunningham <nigelc@catalyst-au.net>